...Implemented restful interface and rewrote getProject api to work with it and it seems to work
