+++ api/restEndpoints/getProject.php
@@ -1,16 +1,8 @@
 <? 
-    require "../../includes/init.inc.php";
     $pdo = get_db();
-
-    //important to tell your browser what we will be sending
-    header("Access-Control-Allow-Origin: *");
-    header("Content-Type: application/json; charset=UTF-8");
-    header("Access-Control-Allow-Methods: GET");
-    header("Access-Control-Allow-Headers: Content-Type, Access-Control-Allow-Headers, Authorization, X-Requested-With");
-
-
+    $reqProjectId = (int)$args[0];
     $returnProject = ['Success' => []];
-    array_push($returnProject['Success'], Array('userType' => 'guest'));
+    $returnProject['Success']['userType'] = 'guest';
 
     //This API doesnt require an authorization header containing the JWT token
     //However if it does additional  info of what type of account is trying to access the data will be return
@@ -33,16 +25,16 @@
     $result = $pdo->prepare("select projectCategory, projectBio, projectBudget, projectDeadline, projectCountry, projectLanguage, projectCurrency, dateEntered, startDate from projects where projectId = :projectId");
 
     $result->execute([
-        'projectId' => $_GET['projectId']
+        'projectId' => $reqProjectId
     ]);
 
     if($result->rowCount() > 0){
         foreach($result as $row){
             pushProjectDetails($returnProject['Success'], $row);
         }
-        echo json_encode($returnProject);
+        return $returnProject;
     }else{
-        echo json_encode(Array('Error' => "Oops, the project you're looking doesn't seem to exist"));
+        return Array('Error' => "Oops, the project you're looking doesn't seem to exist");
     }
 
     function pushProjectDetails(&$returnProject, $info){
