--- api/v1/index.php
@@ -1,3 +1,0 @@
-<?php 
-echo 'Ade'
-?>
