--- views/testView.php
@@ -1,10 +1,0 @@
-<?php 
-    require "../includes/classesLoader.inc.php";
-
-    $thisJWT = new Jwt('');
-    echo $thisJWT->getToken();
-    $thisJWT->setToken(Array('Ade' => 'Akingbade'));
-    $test = $thisJWT->getToken();
-    echo $test;
-?>
-
